move utilites to own module
