+++ utilities/fileExists_test.go
